💄 style(Hooks):
useDelete, useForm, useGet

by @ehfdl
