--- src/hooks/useTodo.jsx
@@ -1,26 +1,0 @@
-import { createAsyncThunk } from "@reduxjs/toolkit";
-import axios from "axios";
-
-export const __getTodos = createAsyncThunk(
-  "GET_TODOS",
-  async (payload, thunkAPI) => {
-    try {
-      const data = await axios.get("http://localhost:3001/todos");
-      return thunkAPI.fulfillWithValue(data.data);
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error);
-    }
-  }
-);
-
-export const __addTodo = createAsyncThunk(
-  "ADD_TODO",
-  async (payload, thunkAPI) => {
-    try {
-      const { data } = await axios.post("http://localhost:3001/todos", payload);
-      return thunkAPI.fulfillWithValue(data);
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error);
-    }
-  }
-);
