+++ src/redux/module/todosSlice.jsx
@@ -1,11 +1,46 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { __getTodos, __addTodo, __deleteTodo } from "../../hooks/useTodo";
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import axios from "axios";
 
 const initialState = {
   todos: [],
   isLoading: false,
   error: null,
 };
+export const __getTodos = createAsyncThunk(
+  "GET_TODOS",
+  async (payload, thunkAPI) => {
+    try {
+      const data = await axios.get("http://localhost:3001/todos");
+      return thunkAPI.fulfillWithValue(data.data);
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error);
+    }
+  }
+);
+export const __addTodo = createAsyncThunk(
+  "ADD_TODO",
+  async (payload, thunkAPI) => {
+    try {
+      const { data } = await axios.post("http://localhost:3001/todos", payload);
+      return thunkAPI.fulfillWithValue(data);
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error);
+    }
+  }
+);
+
+export const __deleteTodo = createAsyncThunk(
+  "DELETE_TODO",
+  async (payload, thunkAPI) => {
+    try {
+      await axios.delete(`http://localhost:3001/todos/${payload}`);
+      return thunkAPI.fulfillWithValue(payload);
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error);
+    }
+  }
+);
 
 export const todosSlice = createSlice({
   name: "todos",
